2407: Address PR requests - Renamed micro-service -> microService - Renamed tsconfig.dependency files - Updated README.md
