+++ packages/entity-server/src/@types/express/index.d.ts
@@ -3,7 +3,6 @@
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 import { AccessPolicy } from '@tupaia/access-policy';
-import { Context } from '@tupaia/server-boilerplate';
 import { EntityServerModelRegistry } from '../../types';
 
 declare global {
@@ -11,11 +10,11 @@ declare global {
     export interface Request {
       accessPolicy: AccessPolicy;
       models: EntityServerModelRegistry;
-      ctx: Context;
+      ctx: Record<string, unknown>;
     }
 
     export interface Response {
-      ctx: Context;
+      ctx: Record<string, unknown>;
     }
   }
 }
