+++ packages/entity-server/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "../../tsconfig-ts.dependency.json",
+  "extends": "../../tsconfig-ts.json",
 
   "compilerOptions": {
     "typeRoots": ["src/@types", "node_modules/@types", "../../node_modules/@types"],
     "outDir": "dist"
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "src/@types"]
+
+  "exclude": ["node_modules", "dist", "src/@types"]
 }
