+++ packages/lesmis-server/tsconfig.json
@@ -1,3 +1,9 @@
 {
-  "extends": "../../tsconfig-ts.dependency.json"
+  "extends": "../../tsconfig-ts.json",
+
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+
+  "exclude": ["node_modules", "dist"]
 }
