+++ packages/server-boilerplate/src/@types/express/index.d.ts
@@ -2,12 +2,10 @@
  * Tupaia
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
-import { LesmisSessionType } from '../../models';
 
 declare global {
   namespace Express {
     export interface Request {
-      session: LesmisSessionType;
       ctx: Record<string, unknown>;
     }
 
@@ -16,3 +14,5 @@ declare global {
     }
   }
 }
+
+export {};
