+++ packages/server-boilerplate/src/index.ts
@@ -7,8 +7,7 @@
 export { ApiConnection } from './connections';
 export { Route } from './routes';
 export { handleWith } from './utils';
-export { Context } from './types';
-export { ApiBuilder as MicroServiceApiBuilder } from './micro-service';
+export { ApiBuilder as MicroServiceApiBuilder } from './microService';
 export {
   ApiBuilder as OrchestratorApiBuilder,
   SessionModel,
