+++ packages/server-boilerplate/src/microService/@types/express/index.d.ts
@@ -4,18 +4,12 @@
  */
 import { AccessPolicy } from '@tupaia/access-policy';
 import { ModelRegistry } from '@tupaia/database';
-import { Context } from '../../../types';
 
 declare global {
   namespace Express {
     export interface Request {
       accessPolicy: AccessPolicy;
       models: ModelRegistry;
-      ctx: Context;
-    }
-
-    export interface Response {
-      ctx: Context;
     }
   }
 }
