+++ packages/server-boilerplate/src/orchestrator/@types/express/index.d.ts
@@ -2,7 +2,6 @@
  * Tupaia
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
-import { Context } from '../../../types';
 import { SessionCookie } from '../../types';
 import { SessionModel, SessionType } from '../../models';
 
@@ -12,11 +11,6 @@ declare global {
       sessionModel: SessionModel;
       sessionCookie?: SessionCookie;
       session?: SessionType;
-      ctx: Context;
-    }
-
-    export interface Response {
-      ctx: Context;
     }
   }
 }
