+++ packages/server-boilerplate/src/routes/Route.ts
@@ -11,13 +11,13 @@ export type Params<Type> = Type extends Request<infer P, any, any, any> ? P : ne
 export type ResBody<Type> = Type extends Request<any, infer sBody, any, any> ? sBody : never;
 export type ReqBody<Type> = Type extends Request<any, any, infer qBody, any> ? qBody : never;
 export type Query<Type> = Type extends Request<any, any, any, infer Q> ? Q : never;
-export type MatchingRequest<Req> = Request<Params<Req>, ResBody<Req>, ReqBody<Req>, Query<Req>>;
-export type MatchingResponse<Req> = Response<ResBody<Req>>;
+export type ExpressRequest<Req> = Request<Params<Req>, ResBody<Req>, ReqBody<Req>, Query<Req>>;
+export type ExpressResponse<Req> = Response<ResBody<Req>>;
 /* eslint-enable @typescript-eslint/no-explicit-any */
 
 export class Route<
-  Req extends MatchingRequest<Req> = Request,
-  Res extends MatchingResponse<Req> = Response
+  Req extends ExpressRequest<Req> = Request,
+  Res extends ExpressResponse<Req> = Response
 > {
   readonly req: Req;
