+++ packages/server-boilerplate/src/types.ts
@@ -6,7 +6,3 @@
 export type AccessPolicyObject = Record<string, string[]>;
 
 export type EmptyObject = Record<string, never>;
-
-export type Context<T> = {
-  [field in keyof T]: T[field];
-};
