+++ packages/utils/tsconfig.json
@@ -1,9 +1,8 @@
 {
-  "extends": "../../tsconfig.dependency.json",
+  "extends": "../../tsconfig-js.json",
   "compilerOptions": {
     "outDir": "dist"
   },
   "include": ["src/**/*.js"],
-  "exclude": [
-    "src/__tests__", "src/migrations"]
+  "exclude": ["src/__tests__", "src/migrations"]
 }
