+++ tsconfig-ts.json
@@ -14,7 +14,5 @@
     "paths": {
       "@tupaia/*": ["*/src"]
     }
-  },
-
-  "exclude": ["node_modules", "dist"]
+  }
 }
